Template -> Bracket Template was too generic Minor code clean up
